Fix adding the runtime user
The approach using the image file and the extrausers class didn't work
because the users didn't exist by the time the do_install here was
called.

The limitation here is that I couldn't find a way to add more than one
user per recipe.

Short of creating a second recipe I opted to use root as the build user
instead.