Add reset button functionality
The Senic hub has one physical button which is used for rebooting the
device as well as resetting it to factory defaults.

The button is connected to the MCU's GPIO pin 10. When the button is open
(default state) the pin is pulled up to 3.3V. By pressing the button the pin
is pulled to GND.

The button handler daemon is started via an init script and will differentiate
between two types of input:

  short press - when the button is pressed and held for at least 50ms (but no
                longer than 9 seconds), a soft reboot is performed

  long press - when the button is pressed and held between 10 and 30 seconds,
               the supervisor daemon is stopped, bluenet wifi settings are
               cleared and a soft reboot is performed

The daemon will log debugging info to syslog (logs normally land in
/var/log/messages prefixed with "senic_button").

The daemon could be started earlier during boot-up, but this way we would lose
the ability to log debug info to disk. If we choose to do so logging directly
to the TTY serial console would still be an option.