u-boot: Update patch 0002 from mender's append
Mender uses 2 patches on u-boot source. The path 0002 does not apply
because we are in recent u-boot version.
A fix upstream has been available but it is currently not available
on "morty" branch.

On senic recipe, we remove the mender's patch and add the newer patch
taken from mender Master's branch:
https://github.com/mendersoftware/meta-mender/commit/5a46970782d7001683b84a7bff0984636e729efa#diff-cbb1ac6ede925922e707c2b20d8f2538

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>