Switch starting supervisor via systemd
Now that we've switched to using systemd we must also create a service
entry for supervisor, since the existing rc script was no longer being
honoured.

Luckily we only need to do this for supervisor itself and not for any of
the services *it* manages.

Also, eventually we will run mender without systemd, but for now we will
keep it this way, since it's the default behaviour of mender.